+++ ui/src/components/General/Notification/components/NotificationList.tsx
@@ -105,6 +105,7 @@ const NotificationList: React.FC<NotificationListProps> = ({
 				"notification-container"
 			)}
 			onOpenChange={handleRedFlagNotification}
+			open
 		>
 			<Button
 				aria-label="notification-icon"