+++ ui/src/components/InjectedComponent/index.ts
@@ -1,0 +1,6 @@
+export { default as ClipboardButton } from "./ClipboardButton";
+export { default as News } from "./ListItems/News";
+export { default as Resource } from "./ListItems/Resource";
+
+export type { ResourceType } from "./ListItems/Resource";
+export type { NewsType } from "./ListItems/News";